requirements.txt
fastapi~=0.105.0
uvicorn~=0.25.0
fastapi-authlib
python-dotenv~=1.0.0
jinja2
google~=3.0.0
starlette~=0.27.0
Authlib~=1.3.0
itsdangerous
httpx
google-auth
google-api-python-client
python-multipart
google-api-core==2.15.0
google-api-python-client==2.111.0
google-auth==2.25.2
google-auth-httplib2==0.2.0
googleapis-common-protos==1.62.0